refactor(timeline): split into a folder, window rows, defer until shown
Ported from wowsims/mop#1563. timeline.tsx becomes a timeline/ folder -
chart series, tooltip content, constants, the chart view picker and the
windowed row split out of what was a 1,800-line file. Rotation row items
mount only while their x-range is on screen, ApexCharts is imported on
demand, and the timeline and log tabs hold their results until first shown
rather than building into a display:none pane.

Two features this repo has and MoP does not are carried into the new
structure rather than lost with the old file:

- The Show GCD toggle, its GCD strip row and the GCD extension drawn past
  the end of each cast bar. Both now go through WindowedRow, so they are
  built per visible range like every other row rather than all at once.
- The cast-delay bar and its tooltip, which marks how long an auto attack
  was delayed before it landed.

Other TBC-side resolutions:

- No secondary resources here, so the GenericResource name/icon override
  is dropped from the resource series, the resource rows and TimelineConfig.
- No OtherActionDamageAmplifier and no Solar/Lunar energy, so those drop
  out of the action categories and the percentage-resource branch.
- No CombatReplay in this repo, so the replay tab and its wiring stay out.
- The chart picker keeps MoP's radio button group; .timeline-show-gcd is
  restored alongside it.
