+++ ui/core/components/detailed_results/result_component.ts
@@ -12,22 +12,43 @@ export interface ResultComponentConfig {
 	parent: HTMLElement;
 	rootCssClass?: string;
 	resultsEmitter: TypedEvent<SimResultData | null>;
+	deferUntilShown?: boolean;
 }
 
 export abstract class ResultComponent extends Component {
 	lastSimResult: SimResultData | null;
 	private resetCallbacks: (() => void)[] = [];
+	private readonly deferUntilShown: boolean;
+	private tabShown: boolean;
+	private pendingSimResult = false;
 
 	constructor(config: ResultComponentConfig) {
 		super(config.parent, config.rootCssClass || 'result-component');
 		this.lastSimResult = null;
+		this.deferUntilShown = !!config.deferUntilShown;
+		this.tabShown = !this.deferUntilShown;
 
 		config.resultsEmitter.on((_, resultData) => {
 			if (!resultData) return;
 
 			this.lastSimResult = resultData;
-			this.onSimResult(resultData);
+			if (this.tabShown) {
+				this.onSimResult(resultData);
+			} else {
+				this.pendingSimResult = true;
+			}
 		});
+	}
+
+	onTabShown() {
+		this.tabShown = true;
+		if (!this.pendingSimResult) return;
+		this.pendingSimResult = false;
+		if (this.lastSimResult) this.onSimResult(this.lastSimResult);
+	}
+
+	onTabHidden() {
+		if (this.deferUntilShown) this.tabShown = false;
 	}
 
 	hasLastSimResult(): boolean {
