--- ui/core/components/detailed_results/timeline.tsx
@@ -1,1799 +1,0 @@
-import ApexCharts from 'apexcharts';
-import clsx from 'clsx';
-import tippy from 'tippy.js';
-import { ref } from 'tsx-vanilla';
-
-import { CacheHandler } from '../../cache_handler';
-import { OtherAction } from '../../proto/common';
-import { ResourceType } from '../../proto/spell';
-import { ActionId, buffAuraToSpellIdMap, resourceTypeToIcon } from '../../proto_utils/action_id';
-import { AuraUptimeLog, CastLog, DpsLog, ResourceChangedLogGroup, SimLog, ThreatLogGroup } from '../../proto_utils/logs_parser';
-import { resourceNames } from '../../proto_utils/names';
-import { UnitMetrics } from '../../proto_utils/sim_result';
-import { orderedResourceTypes } from '../../proto_utils/utils';
-import { TypedEvent } from '../../typed_event';
-import { bucket, distinct, fragmentToString, maxIndex, stringComparator } from '../../utils';
-import { actionColors } from './color_settings';
-import i18n from '../../../i18n/config';
-import { BooleanPicker } from '../pickers/boolean_picker';
-import { ResultComponent, ResultComponentConfig, SimResultData } from './result_component';
-import { APLActionItemSwap_SwapSet } from '../../proto/apl';
-
-type TooltipHandler = (dataPointIndex: number) => Element;
-
-const dpsColor = '#ed5653';
-const manaColor = '#2E93fA';
-const threatColor = '#b56d07';
-
-const cachedSpellCastIcon = new CacheHandler<HTMLAnchorElement>();
-
-interface TimelineConfig extends ResultComponentConfig {}
-
-interface RotationSlot {
-	key: string;
-	labels: Array<Node>;
-	timeline: Array<Node>;
-	hiddenIdsNodes: Array<Node>;
-	emitter: TypedEvent<void>;
-	resetCallbacks: Array<() => void>;
-	plotOptions: any;
-}
-
-export class Timeline extends ResultComponent {
-	private readonly dpsResourcesPlotElem: HTMLElement;
-	private dpsResourcesPlot: any;
-
-	private readonly rotationPlotElem: HTMLElement;
-	private readonly rotationLabels: HTMLElement;
-	private readonly rotationTimeline: HTMLElement;
-	private readonly rotationHiddenIdsContainer: HTMLElement;
-	private readonly chartPicker: HTMLSelectElement;
-	private showGcd = false;
-	private readonly showGcdChangeEmitter = new TypedEvent<void>();
-
-	private resultData: SimResultData | null;
-	rendered: boolean;
-
-	// A rendered rotation timeline for one (result, filter, chart) key. The DOM
-	// nodes are kept LIVE (moved in and out of the containers, never cloned) so
-	// their tippy instances, click handlers and emitter subscriptions survive a
-	// switch between the current result and a saved reference. Eviction runs the
-	// slot's reset callbacks (tooltip destroy, listener removal).
-	private liveSlot: RotationSlot | null = null;
-	// The most recently stashed slot. One is enough for the current <-> reference
-	// swap, and each parked slot keeps its full tippy instance set alive.
-	private parkedSlot: RotationSlot | null = null;
-
-	// Hidden rows are keyed by section (player '', pet name, target label) plus
-	// action, so ids shared across sections (e.g. main-hand Attack) hide independently.
-	private hiddenIds: Array<{ scope: string; actionId: ActionId }>;
-
-	constructor(config: TimelineConfig) {
-		config.rootCssClass = 'timeline-root';
-		super(config);
-		this.resultData = null;
-		this.rendered = false;
-		this.hiddenIds = [];
-		this.addOnDisposeCallback(() => this.reset());
-
-		const showGcdContainerRef = ref<HTMLDivElement>();
-		this.rootElem.appendChild(
-			<div className="timeline-disclaimer">
-				<div className="d-flex flex-column">
-					<p>
-						<i className="warning fa fa-exclamation-triangle fa-xl me-2"></i>
-						{i18n.t('results_tab.details.timeline.disclaimer')}
-					</p>
-					<p>{i18n.t('results_tab.details.timeline.note')}</p>
-				</div>
-				<div className="timeline-controls">
-					<div className="timeline-show-gcd" ref={showGcdContainerRef} />
-					<select className="timeline-chart-picker form-select">
-						<option className="rotation-option" value="rotation">
-							{i18n.t('results_tab.details.timeline.chart_types.rotation')}
-						</option>
-						<option className="dps-option" value="dps">
-							{i18n.t('results_tab.details.timeline.chart_types.dps')}
-						</option>
-						<option className="threat-option" value="threat">
-							{i18n.t('results_tab.details.timeline.chart_types.threat')}
-						</option>
-					</select>
-				</div>
-			</div>,
-		);
-
-		this.rootElem.appendChild(
-			<div className="timeline-plots-container">
-				<div className="timeline-plot dps-resources-plot hide"></div>
-				<div className="timeline-plot rotation-plot">
-					<div className="rotation-container">
-						<div className="rotation-labels"></div>
-						<div className="rotation-timeline" draggable={true}></div>
-					</div>
-					<div className="rotation-hidden-ids"></div>
-				</div>
-			</div>,
-		);
-
-		this.chartPicker = this.rootElem.querySelector('.timeline-chart-picker')!;
-		this.chartPicker.addEventListener('change', () => this.onChartPickerSelectHandler());
-
-		this.dpsResourcesPlotElem = this.rootElem.querySelector('.dps-resources-plot')!;
-		this.dpsResourcesPlot = new ApexCharts(this.dpsResourcesPlotElem, {
-			chart: {
-				animations: {
-					enabled: false,
-				},
-				background: 'transparent',
-				foreColor: 'white',
-				height: '100%',
-				id: 'dpsResources',
-				type: 'line',
-				zoom: {
-					enabled: true,
-					allowMouseWheelZoom: false,
-				},
-			},
-			series: [], // Set dynamically
-			xaxis: {
-				title: {
-					text: i18n.t('results_tab.details.timeline.chart_options.time_axis'),
-				},
-			},
-			noData: {
-				text: i18n.t('results_tab.details.timeline.chart_options.waiting_for_data'),
-			},
-			stroke: {
-				width: 2,
-				curve: 'straight',
-			},
-		});
-
-		this.rotationPlotElem = this.rootElem.querySelector('.rotation-plot')!;
-		this.rotationLabels = this.rootElem.querySelector('.rotation-labels')!;
-		this.rotationTimeline = this.rootElem.querySelector('.rotation-timeline')!;
-		this.rotationHiddenIdsContainer = this.rootElem.querySelector('.rotation-hidden-ids')!;
-
-		new BooleanPicker<null>(showGcdContainerRef.value!, null, {
-			id: 'timeline-show-gcd',
-			label: i18n.t('results_tab.details.timeline.show_gcd'),
-			inline: true,
-			changedEvent: () => this.showGcdChangeEmitter,
-			getValue: () => this.showGcd,
-			setValue: (eventID, _, newValue) => {
-				this.showGcd = newValue;
-				this.rotationPlotElem.classList.toggle('show-gcd', newValue);
-				this.showGcdChangeEmitter.emit(eventID);
-			},
-		});
-
-		let isMouseDown = false;
-		let startX = 0;
-		let scrollLeft = 0;
-		this.rotationTimeline.addEventListener('dragstart', event => {
-			event.preventDefault();
-		});
-		this.rotationTimeline.addEventListener('mousedown', event => {
-			isMouseDown = true;
-			startX = event.pageX - this.rotationTimeline.offsetLeft;
-			scrollLeft = this.rotationTimeline.scrollLeft;
-		});
-		this.rotationTimeline.addEventListener('mouseleave', () => {
-			isMouseDown = false;
-			this.rotationTimeline.classList.remove('active');
-		});
-		this.rotationTimeline.addEventListener('mouseup', () => {
-			isMouseDown = false;
-			this.rotationTimeline.classList.remove('active');
-		});
-		this.rotationTimeline.addEventListener('mousemove', event => {
-			if (!isMouseDown) return;
-			event.preventDefault();
-			const x = event.pageX - this.rotationTimeline.offsetLeft;
-			const walk = (x - startX) * 3; //scroll-fast
-			this.rotationTimeline.scrollLeft = scrollLeft - walk;
-		});
-	}
-
-	onChartPickerSelectHandler() {
-		if (this.chartPicker.value === 'rotation') {
-			this.dpsResourcesPlotElem.classList.add('hide');
-			this.rotationPlotElem.classList.remove('hide');
-		} else {
-			this.dpsResourcesPlotElem.classList.remove('hide');
-			this.rotationPlotElem.classList.add('hide');
-		}
-	}
-
-	onSimResult(resultData: SimResultData) {
-		this.resultData = resultData;
-		this.update();
-	}
-
-	private updatePlot() {
-		if (this.resultData == null) {
-			return;
-		}
-
-		const players = this.resultData.result.getRaidIndexedPlayers(this.resultData.filter);
-		const singlePlayer = players.length == 1;
-		if (!singlePlayer && this.chartPicker.value == 'rotation') {
-			// Programmatic select changes fire no 'change' event: sync the plot containers by hand.
-			this.chartPicker.value = 'dps';
-			this.onChartPickerSelectHandler();
-		}
-
-		// Fast path: this result was rendered before and its slot is either live
-		// or parked in the cache.
-		const key = this.resultKey(!singlePlayer);
-		const hit = this.liveSlot?.key === key ? this.liveSlot : this.parkedSlot?.key === key ? this.parkedSlot : null;
-		if (hit?.plotOptions) {
-			if (hit !== this.liveSlot) {
-				this.parkedSlot = null;
-				this.stashLiveSlot();
-				this.attachSlot(hit);
-			}
-			this.setRotationOptionVisible(singlePlayer);
-			this.dpsResourcesPlot.updateOptions(hit.plotOptions);
-			return;
-		}
-
-		const duration = this.resultData!.result.result.firstIterationDuration || 1;
-		const options: any = {
-			theme: {
-				mode: 'dark',
-			},
-			series: [],
-			colors: [],
-			xaxis: {
-				min: 0,
-				max: duration,
-				tickAmount: 10,
-				decimalsInFloat: 1,
-				labels: {
-					show: true,
-				},
-				title: {
-					text: 'Time (s)',
-				},
-			},
-			yaxis: [],
-			chart: {
-				events: {
-					beforeResetZoom: () => {
-						return {
-							xaxis: {
-								min: 0,
-								max: duration,
-							},
-						};
-					},
-				},
-				toolbar: {
-					show: false,
-				},
-			},
-		};
-
-		let tooltipHandlers: Array<TooltipHandler | null> = [];
-		options.tooltip = {
-			enabled: true,
-			custom: (data: { series: any; seriesIndex: number; dataPointIndex: number; w: any }) => {
-				if (tooltipHandlers[data.seriesIndex]) {
-					return fragmentToString(tooltipHandlers[data.seriesIndex]!(data.dataPointIndex));
-				} else {
-					throw new Error('No tooltip handler for series ' + data.seriesIndex);
-				}
-			},
-		};
-
-		if (singlePlayer) {
-			const player = players[0];
-
-			this.setRotationOptionVisible(true);
-
-			try {
-				this.updateRotationChart(player, duration);
-			} catch (e) {
-				console.log('Failed to update rotation chart: ', e);
-			}
-
-			const dpsData = this.addDpsSeries(player, options, '');
-			this.addDpsYAxis(dpsData.maxDps, options);
-			tooltipHandlers.push(dpsData.tooltipHandler);
-			tooltipHandlers.push(this.addManaSeries(player, options));
-			tooltipHandlers.push(this.addThreatSeries(player, options, ''));
-			tooltipHandlers = tooltipHandlers.filter(handler => !!handler);
-
-			this.addMajorCooldownAnnotations(player, options);
-		} else {
-			this.setRotationOptionVisible(false);
-
-			this.stashLiveSlot();
-			this.clearRotationChart();
-			// No rotation subtree, but the (expensive) per-player series are worth caching for swaps.
-			this.liveSlot = Timeline.newSlot(key);
-
-			if (this.chartPicker.value == 'dps') {
-				let maxDps = 0;
-				players.forEach(player => {
-					const dpsData = this.addDpsSeries(player, options, `var(--bs-${player.classColor}`);
-					maxDps = Math.max(maxDps, dpsData.maxDps);
-					tooltipHandlers.push(dpsData.tooltipHandler);
-				});
-				this.addDpsYAxis(maxDps, options);
-			} else {
-				// threat
-				let maxThreat = 0;
-				players.forEach(player => {
-					tooltipHandlers.push(this.addThreatSeries(player, options, player.classColor));
-					maxThreat = Math.max(maxThreat, player.maxThreat);
-				});
-				this.addThreatYAxis(maxThreat, options);
-			}
-		}
-
-		if (this.liveSlot?.key === key) this.liveSlot.plotOptions = options;
-		this.dpsResourcesPlot.updateOptions(options);
-	}
-
-	private addDpsYAxis(maxDps: number, options: any) {
-		const dpsAxisMax = Math.ceil(maxDps / 100) * 100;
-		options.yaxis.push({
-			color: dpsColor,
-			seriesName: 'DPS',
-			min: 0,
-			max: dpsAxisMax,
-			tickAmount: 10,
-			decimalsInFloat: 0,
-			title: {
-				text: 'DPS',
-				style: {
-					color: dpsColor,
-				},
-			},
-			axisBorder: {
-				show: true,
-				color: dpsColor,
-			},
-			axisTicks: {
-				color: dpsColor,
-			},
-			labels: {
-				minWidth: 30,
-				style: {
-					colors: [dpsColor],
-				},
-			},
-		});
-	}
-
-	private addThreatYAxis(maxThreat: number, options: any) {
-		const axisMax = Math.ceil(maxThreat / 10000) * 10000;
-		options.yaxis.push({
-			color: threatColor,
-			seriesName: i18n.t('results_tab.details.timeline.tooltips.threat'),
-			min: 0,
-			max: axisMax,
-			tickAmount: 10,
-			decimalsInFloat: 0,
-			title: {
-				text: i18n.t('results_tab.details.timeline.tooltips.threat'),
-				style: {
-					color: threatColor,
-				},
-			},
-			axisBorder: {
-				show: true,
-				color: threatColor,
-			},
-			axisTicks: {
-				color: threatColor,
-			},
-			labels: {
-				minWidth: 30,
-				style: {
-					colors: [threatColor],
-				},
-			},
-		});
-	}
-
-	// Returns a function for drawing the tooltip, or null if no series was added.
-	private addDpsSeries(unit: UnitMetrics, options: any, colorOverride: string): { maxDps: number; tooltipHandler: TooltipHandler } {
-		const dpsLogs = unit.dpsLogs.filter(log => log.timestamp >= 0);
-
-		options.colors.push(colorOverride || dpsColor);
-		options.series.push({
-			name: 'DPS',
-			type: 'line',
-			data: dpsLogs.map(log => {
-				return {
-					x: log.timestamp,
-					y: log.dps,
-				};
-			}),
-		});
-
-		return {
-			maxDps: dpsLogs[maxIndex(dpsLogs.map(l => l.dps))!]?.dps,
-			tooltipHandler: (dataPointIndex: number) => {
-				const log = dpsLogs[dataPointIndex];
-				return this.dpsTooltip(log, true, unit, colorOverride);
-			},
-		};
-	}
-
-	// Returns a function for drawing the tooltip, or null if no series was added.
-	private addManaSeries(unit: UnitMetrics, options: any): TooltipHandler | null {
-		const manaLogs = unit.groupedResourceLogs[ResourceType.ResourceTypeMana].filter(log => log.timestamp >= 0);
-		if (manaLogs.length == 0) {
-			return null;
-		}
-		const maxMana = manaLogs[0].valueBefore;
-
-		options.colors.push(manaColor);
-		options.series.push({
-			name: 'Mana',
-			type: 'line',
-			data: manaLogs.map(log => {
-				return {
-					x: log.timestamp,
-					y: log.valueAfter,
-				};
-			}),
-		});
-		options.yaxis.push({
-			seriesName: 'Mana',
-			opposite: true, // Appear on right side
-			min: 0,
-			max: maxMana,
-			tickAmount: 10,
-			title: {
-				text: 'Mana',
-				style: {
-					color: manaColor,
-				},
-			},
-			axisBorder: {
-				show: true,
-				color: manaColor,
-			},
-			axisTicks: {
-				color: manaColor,
-			},
-			labels: {
-				minWidth: 30,
-				style: {
-					colors: [manaColor],
-				},
-				formatter: (val: string) => {
-					const v = parseFloat(val);
-					return `${v.toFixed(0)} (${((v / maxMana) * 100).toFixed(0)}%)`;
-				},
-			},
-		} as any);
-
-		return (dataPointIndex: number) => {
-			const log = manaLogs[dataPointIndex];
-			return this.resourceTooltip(log, maxMana, true);
-		};
-	}
-
-	// Returns a function for drawing the tooltip, or null if no series was added.
-	private addThreatSeries(unit: UnitMetrics, options: any, colorOverride: string): TooltipHandler | null {
-		options.colors.push(colorOverride || threatColor);
-		options.series.push({
-			name: i18n.t('results_tab.details.timeline.tooltips.threat'),
-			type: 'line',
-			data: unit.threatLogs
-				.filter(log => log.timestamp >= 0)
-				.map(log => {
-					return {
-						x: log.timestamp,
-						y: log.threatAfter,
-					};
-				}),
-		});
-
-		return (dataPointIndex: number) => {
-			const log = unit.threatLogs[dataPointIndex];
-			return this.threatTooltip(log, true, unit, colorOverride);
-		};
-	}
-
-	private addMajorCooldownAnnotations(unit: UnitMetrics, options: any) {
-		const mcdLogs = unit.majorCooldownLogs;
-		const mcdAuraLogs = unit.majorCooldownAuraUptimeLogs;
-
-		// Figure out how much to vertically offset cooldown icons, for cooldowns
-		// used very close to each other. This is so the icons don't overlap.
-		const MAX_ALLOWED_DIST = 10;
-		const cooldownIconOffsets = mcdLogs.map(
-			(mcdLog, mcdIdx) => mcdLogs.filter((cdLog, cdIdx) => cdIdx < mcdIdx && cdLog.timestamp > mcdLog.timestamp - MAX_ALLOWED_DIST).length,
-		);
-
-		const distinctMcdAuras = distinct(mcdAuraLogs, (a, b) => a.actionId!.equalsIgnoringTag(b.actionId!));
-		// Sort by name so auras keep their same colors even if timings change.
-		distinctMcdAuras.sort((a, b) => stringComparator(a.actionId!.name, b.actionId!.name));
-		const mcdAuraColors = mcdAuraLogs.map(
-			mcdAuraLog => actionColors[distinctMcdAuras.findIndex(dAura => dAura.actionId!.equalsIgnoringTag(mcdAuraLog.actionId!))],
-		);
-
-		options.annotations = {
-			position: 'back',
-			xaxis: mcdAuraLogs.map((log, i) => {
-				return {
-					x: log.gainedAt,
-					x2: log.fadedAt,
-					fillColor: mcdAuraColors[i],
-				};
-			}),
-			points: mcdLogs.map((log, i) => {
-				return {
-					x: log.timestamp,
-					y: 0,
-					image: {
-						path: log.actionId!.iconUrl,
-						width: 20,
-						height: 20,
-						offsetY: cooldownIconOffsets[i] * -25,
-					},
-				};
-			}),
-		};
-	}
-
-	private clearRotationChart() {
-		this.rotationLabels.replaceChildren(<div className="rotation-label-header"></div>);
-		this.rotationTimeline.replaceChildren(
-			<div className="rotation-timeline-header">
-				<canvas className="rotation-timeline-canvas" />
-			</div>,
-		);
-		this.rotationHiddenIdsContainer.replaceChildren();
-	}
-
-	private updateRotationChart(player: UnitMetrics, duration: number) {
-		const targets = this.resultData!.result.getTargets(this.resultData!.filter);
-		if (targets.length == 0) {
-			return;
-		}
-
-		const key = this.resultKey();
-		if (this.liveSlot?.key === key) {
-			return;
-		}
-		// Take before stashing so the stash can't evict the slot we want.
-		const parked = this.takeParkedSlot(key);
-		this.stashLiveSlot();
-		if (parked) {
-			this.attachSlot(parked);
-			return;
-		}
-		this.liveSlot = Timeline.newSlot(key);
-		this.clearRotationChart();
-
-		try {
-			this.drawRotationTimeRuler(this.rotationTimeline.querySelector('.rotation-timeline-canvas')!, duration);
-		} catch (e) {
-			console.log('Failed to draw rotation: ', e);
-		}
-
-		orderedResourceTypes.forEach(resourceType => this.addResourceRow(resourceType, player.groupedResourceLogs[resourceType], duration));
-
-		const buffsById = Object.values(bucket(player.auraUptimeLogs, log => log.actionId!.toString()));
-		buffsById.sort((a, b) => stringComparator(a[0].actionId!.name, b[0].actionId!.name));
-		const debuffsByTargetById = targets.map(target =>
-			Object.values(bucket(target.auraUptimeLogs, log => log.actionId!.toString())).sort((a, b) =>
-				stringComparator(a[0].actionId!.name, b[0].actionId!.name),
-			),
-		);
-
-		const buffsAndDebuffsById = buffsById.concat(
-			// Only pick target 0 to prevent overlapping cast rows
-			debuffsByTargetById[0],
-		);
-
-		auraAsResource.forEach(actionId => {
-			const auraIndex = buffsById.findIndex(auraUptimeLogs => auraUptimeLogs?.[0].actionId!.equals(actionId));
-			if (auraIndex !== -1) {
-				this.addAuraRow(buffsById[auraIndex], duration);
-			}
-		});
-
-		this.addGcdStripRow(player, duration);
-
-		const playerCastsByAbility = this.getSortedCastsByAbility(player);
-		playerCastsByAbility.forEach(castLogs => this.addCastRow(castLogs, buffsAndDebuffsById, duration));
-
-		if (player.pets.length > 0) {
-			const playerPets = new Map<string, UnitMetrics>();
-			player.pets.forEach(petsLog => {
-				const petCastsByAbility = this.getSortedCastsByAbility(petsLog);
-				if (petCastsByAbility.length > 0) {
-					// Because multiple pets can have the same name and we parse cast logs
-					// by pet name each individual pet ends up with all the casts of pets
-					// with the same name. Because of this we can just grab the first pet
-					// of each name and visualize only that.
-					if (!playerPets.has(petsLog.name)) {
-						playerPets.set(petsLog.name, petsLog);
-					}
-				}
-			});
-
-			playerPets.forEach(pet => {
-				this.addSeparatorRow(duration);
-				this.addPetRow(pet.name, duration);
-				orderedResourceTypes.forEach(resourceType => this.addResourceRow(resourceType, pet.groupedResourceLogs[resourceType], duration));
-				this.addGcdStripRow(pet, duration);
-				const petCastsByAbility = this.getSortedCastsByAbility(pet);
-				petCastsByAbility.forEach(castLogs => this.addCastRow(castLogs, buffsAndDebuffsById, duration, pet.name));
-			});
-		}
-
-		// Don't add a row for buffs that were already visualized in a cast row or are prioritized.
-		const buffsToShow = buffsById.filter(auraUptimeLogs =>
-			playerCastsByAbility.findIndex(casts => {
-				const actionId = auraUptimeLogs[0].actionId;
-				return actionId && (casts[0].actionId!.equalsIgnoringTag(actionId) || auraAsResource.find(auraId => auraId.equals(actionId)));
-			}),
-		);
-		if (buffsToShow.length > 0) {
-			this.addSeparatorRow(duration);
-			buffsToShow.forEach(auraUptimeLogs => this.addAuraRow(auraUptimeLogs, duration));
-		}
-
-		targets.forEach(target => {
-			const targetCastsByAbility = this.getSortedCastsByAbility(target);
-			if (targetCastsByAbility.length > 0) {
-				this.addSeparatorRow(duration);
-				this.addTargetRow(target.label, duration);
-				targetCastsByAbility.forEach(castLogs => this.addCastRow(castLogs, buffsAndDebuffsById, duration, target.label));
-			}
-		});
-
-		// Add a row for all debuffs, even those which have already been visualized in a cast row.
-		debuffsByTargetById.forEach((debuffsToShow, index) => {
-			if (debuffsToShow.length > 0) {
-				this.addSeparatorRow(duration);
-				this.addTargetRow(targets?.[index]?.label, duration);
-				debuffsToShow.forEach(auraUptimeLogs => this.addAuraRow(auraUptimeLogs, duration, targets?.[index]?.label ?? ''));
-			}
-		});
-	}
-
-	private getSortedCastsByAbility(player: UnitMetrics): Array<Array<CastLog>> {
-		const meleeActionIds = player.getMeleeActions().map(action => action.actionId);
-		const spellActionIds = player.getSpellActions().map(action => action.actionId);
-		const getActionCategory = (actionId: ActionId): number => {
-			const fixedCategory = idToCategoryMap[actionId.anyId()];
-			if (fixedCategory != null) {
-				return fixedCategory;
-			} else if (meleeActionIds.find(meleeActionId => meleeActionId.equals(actionId))) {
-				return MELEE_ACTION_CATEGORY;
-			} else if (spellActionIds.find(spellActionId => spellActionId.equals(actionId))) {
-				return SPELL_ACTION_CATEGORY;
-			} else {
-				return DEFAULT_ACTION_CATEGORY;
-			}
-		};
-
-		const castsByAbility = Object.values(
-			bucket(player.castLogs, log => {
-				if (idsToGroupForRotation.includes(log.actionId!.spellId)) {
-					return log.actionId!.toStringIgnoringTag();
-				} else {
-					return log.actionId!.toString();
-				}
-			}),
-		);
-
-		castsByAbility.sort((a, b) => {
-			const categoryA = getActionCategory(a[0].actionId!);
-			const categoryB = getActionCategory(b[0].actionId!);
-			if (categoryA != categoryB) {
-				return categoryA - categoryB;
-			} else if (a[0].actionId!.anyId() == b[0].actionId!.anyId()) {
-				return a[0].actionId!.tag - b[0].actionId!.tag;
-			} else {
-				return stringComparator(a[0].actionId!.name, b[0].actionId!.name);
-			}
-		});
-
-		return castsByAbility;
-	}
-
-	private hiddenIndex(scope: string, actionId: ActionId): number {
-		return this.hiddenIds.findIndex(hidden => hidden.scope === scope && hidden.actionId.equals(actionId));
-	}
-
-	private makeLabelElem(actionId: ActionId, isHiddenLabel: boolean, isAura: boolean, scope: string): JSX.Element {
-		const baseText = idsToGroupForRotation.includes(actionId.spellId) ? actionId.baseName : actionId.name;
-		// Hidden chips for pet/target rows name their section so two "Attack" chips are telling apart.
-		const labelText = isHiddenLabel && scope ? `${baseText} (${scope})` : baseText;
-		const labelIcon = ref<HTMLAnchorElement>();
-		const hideElem = ref<HTMLElement>();
-		const labelElem = (
-			<div className={clsx('rotation-label rotation-row', isHiddenLabel && 'rotation-label-hidden')}>
-				<span ref={hideElem} className={clsx('fas', isHiddenLabel ? 'fa-eye' : 'fa-eye-slash')}></span>
-				<a ref={labelIcon} className="rotation-label-icon"></a>
-				<span className="rotation-label-text">{labelText}</span>
-			</div>
-		);
-		// The whole hidden chip un-hides; on the visible label only the eye hides.
-		const clickTarget = isHiddenLabel ? labelElem : hideElem.value!;
-		const onClickHandler = (event: Event) => {
-			// The spell icon keeps its own link + Wowhead tooltip.
-			if (isHiddenLabel && (event.target as Element).closest('.rotation-label-icon')) return;
-			if (isHiddenLabel) {
-				const index = this.hiddenIndex(scope, actionId);
-				if (index != -1) {
-					this.hiddenIds.splice(index, 1);
-				}
-			} else {
-				this.hiddenIds.push({ scope, actionId });
-			}
-			this.liveSlot?.emitter.emit(TypedEvent.nextEventID());
-		};
-		clickTarget.addEventListener('click', onClickHandler);
-		const tooltip = tippy(hideElem.value!, {
-			theme: 'timeline-tooltip',
-			placement: 'bottom',
-			content: isHiddenLabel ? 'Show Row' : 'Hide Row',
-		});
-
-		const updateHidden = () => {
-			if (isHiddenLabel == (this.hiddenIndex(scope, actionId) != -1)) {
-				labelElem.classList.remove('hide');
-			} else {
-				labelElem.classList.add('hide');
-			}
-		};
-		const event = this.liveSlot!.emitter.on(updateHidden);
-		updateHidden();
-		actionId.setBackgroundAndHref(labelIcon.value!);
-		actionId.setWowheadDataset(labelIcon.value!, { useBuffAura: isAura });
-
-		this.addOnResetCallback(() => {
-			clickTarget.removeEventListener('click', onClickHandler);
-			tooltip.destroy();
-			event.dispose();
-		});
-
-		return labelElem;
-	}
-
-	// A timeline row that is never hidden (section headers: pet name, target name).
-	private makePlainRowElem(duration: number): JSX.Element {
-		return (
-			<div
-				className="rotation-timeline-row rotation-row"
-				style={{
-					width: this.timeToPx(duration),
-				}}></div>
-		);
-	}
-
-	private makeRowElem(actionId: ActionId, duration: number, scope: string): JSX.Element {
-		const rowElem = this.makePlainRowElem(duration);
-
-		const updateHidden = () => {
-			if (this.hiddenIndex(scope, actionId) != -1) {
-				rowElem.classList.add('hide');
-			} else {
-				rowElem.classList.remove('hide');
-			}
-		};
-		const event = this.liveSlot!.emitter.on(updateHidden);
-		updateHidden();
-		this.addOnResetCallback(() => event.dispose());
-		return rowElem;
-	}
-
-	private addPetRow(petName: string, duration: number) {
-		const actionId = ActionId.fromPetName(petName);
-		// Header row: must not follow hiddenIds. fromPetName resolves to the summon
-		// spell's id, so hiding e.g. the "Dire Beast" cast row used to hide this
-		// row but not its label, shifting every row below it.
-		const rowElem = this.makePlainRowElem(duration);
-
-		const iconElem = document.createElement('div');
-		this.rotationLabels.appendChild(iconElem);
-
-		actionId.fill().then(filledActionId => {
-			const labelText = idsToGroupForRotation.includes(filledActionId.spellId) ? filledActionId.baseName : filledActionId.name;
-			const labelIcon = ref<HTMLAnchorElement>();
-			const labelElem = (
-				<div className="rotation-label rotation-row">
-					<a ref={labelIcon} className="rotation-label-icon"></a>
-					<span className="rotation-label-text">{labelText}</span>
-				</div>
-			);
-			filledActionId.setBackgroundAndHref(labelIcon.value!);
-			iconElem.appendChild(labelElem);
-		});
-
-		this.rotationTimeline.appendChild(rowElem);
-	}
-
-	private addTargetRow(targetName: string, duration: number) {
-		const rowElem = this.makePlainRowElem(duration);
-		this.rotationLabels.appendChild(
-			<div>
-				<div className="rotation-label rotation-row">
-					<span className="rotation-label-text">{targetName}</span>
-				</div>
-			</div>,
-		);
-		this.rotationTimeline.appendChild(rowElem);
-	}
-
-	private addGcdStripRow(player: UnitMetrics, duration: number) {
-		const gcdCasts = player.castLogs.filter(c => c.gcd > 0 && !c.castCancelledLog).sort((a, b) => a.timestamp - b.timestamp);
-		if (gcdCasts.length === 0) return;
-
-		this.rotationLabels.appendChild(
-			<div className="rotation-label rotation-row" dataset={{ row: 'gcd' }}>
-				<a className="rotation-label-icon rotation-label-icon-empty" />
-				<span className="rotation-label-text">GCD</span>
-			</div>,
-		);
-
-		const rowElem = (
-			<div
-				className="rotation-timeline-row rotation-row rotation-timeline-gcd-strip"
-				dataset={{ row: 'gcd' }}
-				style={{ width: this.timeToPx(duration) }}
-			/>
-		) as HTMLElement;
-
-		gcdCasts.forEach(c => {
-			const visibleGcd = Math.min(c.gcd, duration - c.timestamp);
-			if (visibleGcd <= 0) return;
-
-			const segElem = (
-				<div
-					className="rotation-timeline-gcd-segment"
-					style={{
-						left: this.timeToPx(c.timestamp),
-						width: this.timeToPx(visibleGcd),
-					}}
-				/>
-			) as HTMLElement;
-			rowElem.appendChild(segElem);
-
-			const segTip = tippy(segElem, {
-				placement: 'bottom',
-				content: (
-					<div className="timeline-tooltip">
-						<span>
-							{c.actionId!.name} — {c.gcd.toFixed(2)}s GCD ({c.timestamp.toFixed(2)}s → {(c.timestamp + c.gcd).toFixed(2)}s)
-						</span>
-					</div>
-				),
-			});
-			this.addOnResetCallback(() => segTip.destroy());
-		});
-
-		this.rotationTimeline.appendChild(rowElem);
-	}
-
-	private addSeparatorRow(duration: number) {
-		const separatorElem = <div className="rotation-timeline-separator"></div>;
-		this.rotationLabels.appendChild(separatorElem.cloneNode());
-		separatorElem.style.width = this.timeToPx(duration);
-		this.rotationTimeline.appendChild(separatorElem);
-	}
-
-	private addResourceRow(resourceType: ResourceType, resourceLogs: Array<ResourceChangedLogGroup>, duration: number) {
-		if (resourceLogs.length == 0) {
-			return;
-		}
-		const startValue = function (group: ResourceChangedLogGroup): number {
-			if (group.maxValue == null) {
-				return resourceLogs[0].valueBefore;
-			}
-
-			return group.maxValue;
-		};
-
-		let resourceName = resourceNames.get(resourceType);
-		let resourceIcon = resourceTypeToIcon[resourceType];
-
-		const labelElem = (
-			<div className="rotation-label rotation-row">
-				<a
-					className="rotation-label-icon"
-					style={{
-						backgroundImage: `url('${resourceIcon}')`,
-					}}></a>
-				<span className="rotation-label-text">{resourceName}</span>
-			</div>
-		);
-
-		this.rotationLabels.appendChild(labelElem);
-
-		const rowElem = (
-			<div
-				className="rotation-timeline-row rotation-row"
-				style={{
-					width: this.timeToPx(duration),
-				}}></div>
-		);
-
-		resourceLogs.forEach((resourceLogGroup, i) => {
-			const cNames = resourceNames.get(resourceType)!.toLowerCase().replaceAll(' ', '-');
-			const resourceElem = (
-				<div
-					className={`rotation-timeline-resource series-color ${cNames}`}
-					style={{
-						left: this.timeToPx(resourceLogGroup.timestamp),
-						width: this.timeToPx((resourceLogs[i + 1]?.timestamp || duration) - resourceLogGroup.timestamp),
-					}}></div>
-			);
-
-			if (percentageResources.includes(resourceType)) {
-				resourceElem.textContent = ((resourceLogGroup.valueAfter / startValue(resourceLogGroup)) * 100).toFixed(0) + '%';
-			} else {
-				if (resourceType == ResourceType.ResourceTypeEnergy || resourceType == ResourceType.ResourceTypeFocus) {
-					const bgElem = document.createElement('div');
-					bgElem.classList.add('rotation-timeline-resource-fill');
-					bgElem.classList.add(cNames);
-					bgElem.style.height = ((resourceLogGroup.valueAfter / startValue(resourceLogGroup)) * 100).toFixed(0) + '%';
-					resourceElem.appendChild(bgElem);
-				} else {
-					resourceElem.textContent = Math.floor(resourceLogGroup.valueAfter).toFixed(0);
-				}
-			}
-			rowElem.appendChild(resourceElem);
-
-			const tooltip = tippy(resourceElem, {
-				placement: 'bottom',
-				content: this.resourceTooltipElem(resourceLogGroup, startValue(resourceLogGroup), false),
-			});
-			this.addOnResetCallback(() => tooltip.destroy());
-		});
-		this.rotationTimeline.appendChild(rowElem);
-	}
-
-	private addCastRow(castLogs: Array<CastLog>, aurasById: Array<Array<AuraUptimeLog>>, duration: number, scope = '') {
-		const actionId = castLogs[0].actionId!;
-
-		this.rotationLabels.appendChild(this.makeLabelElem(actionId, false, false, scope));
-		this.rotationHiddenIdsContainer.appendChild(this.makeLabelElem(actionId, true, false, scope));
-
-		const rowElem = this.makeRowElem(actionId, duration, scope);
-		castLogs.forEach(castLog => {
-			if (castLog.delay > 0) {
-				const delayElem = (
-					<div
-						className="rotation-timeline-cast-delay"
-						style={{
-							left: this.timeToPx(Math.max(0, castLog.timestamp - castLog.delay)),
-							width: this.timeToPx(castLog.delay),
-						}}
-					/>
-				) as HTMLElement;
-				rowElem.appendChild(delayElem);
-
-				const delayTooltip = tippy(delayElem, {
-					placement: 'bottom',
-					content: (
-						<div className="timeline-tooltip">
-							<span>
-								Auto delayed by {castLog.delayText}, was ready at {castLog.readyAtText}
-							</span>
-						</div>
-					),
-				});
-				this.addOnResetCallback(() => delayTooltip.destroy());
-			}
-
-			const castElem = (
-				<div
-					className="rotation-timeline-cast"
-					style={{
-						left: this.timeToPx(castLog.timestamp),
-						minWidth: this.timeToPx(castLog.cancelTime || castLog.castTime + castLog.travelTime),
-					}}
-				/>
-			);
-			rowElem.appendChild(castElem);
-
-			const extensionStart = castLog.timestamp + castLog.castTime;
-			const gcdOverhead = Math.min(castLog.timestamp + castLog.gcd, duration) - extensionStart;
-			if (gcdOverhead > 0 && !castLog.castCancelledLog) {
-				const gcdExtensionElem = (
-					<div
-						className="rotation-timeline-gcd-extension"
-						style={{
-							left: this.timeToPx(extensionStart),
-							width: this.timeToPx(gcdOverhead),
-						}}
-					/>
-				) as HTMLElement;
-				rowElem.appendChild(gcdExtensionElem);
-			}
-
-			if (castLog.cancelTime) {
-				castElem.classList.add('cast-cancelled');
-			} else if (castLog.travelTime != 0) {
-				const travelTimeElem = (
-					<div
-						className="rotation-timeline-travel-time"
-						style={{
-							left: this.timeToPx(castLog.castTime),
-							minWidth: this.timeToPx(castLog.travelTime),
-						}}
-					/>
-				);
-				castElem.appendChild(travelTimeElem);
-			}
-
-			if (castLog.damageDealtLogs.length > 0) {
-				const ddl = castLog.damageDealtLogs[0];
-				if (ddl.miss || ddl.dodge || ddl.parry) {
-					castElem.classList.add('outcome-miss');
-				} else if (ddl.glance || ddl.block || ddl.partialResist1_4 || ddl.partialResist2_4 || ddl.partialResist3_4) {
-					castElem.classList.add('outcome-partial');
-				} else if (ddl.crit) {
-					castElem.classList.add('outcome-crit');
-				} else {
-					castElem.classList.add('outcome-hit');
-				}
-			}
-
-			const actionIdAsString = actionId.toString();
-			const cachedIconElem = cachedSpellCastIcon.get(actionIdAsString)?.cloneNode() as HTMLAnchorElement | undefined;
-			let iconElem = cachedIconElem;
-			if (!iconElem) {
-				iconElem = (<a className="rotation-timeline-cast-icon" />) as HTMLAnchorElement;
-				actionId.setBackground(iconElem);
-				cachedSpellCastIcon.set(actionIdAsString, iconElem);
-			}
-			castElem.appendChild(iconElem);
-
-			const travelTimeStr = castLog.travelTime == 0 ? '' : ` + ${castLog.travelTime.toFixed(2)}s travel time`;
-			const totalDamage = castLog.totalDamage();
-
-			let timingStr = '';
-			if (castLog.castCancelledLog?.timestamp) {
-				timingStr = ` (Cancelled after ${castLog.cancelTime.toFixed(2)}s)`;
-			} else {
-				const parts: string[] = [];
-				if (castLog.castTime > 0) parts.push(`${castLog.castTime.toFixed(2)}s cast`);
-				if (castLog.gcd > 0) parts.push(`${castLog.gcd.toFixed(2)}s GCD`);
-				if (parts.length > 0) timingStr = ` (${parts.join(', ')})`;
-			}
-
-			const tt = (
-				<div className="timeline-tooltip">
-					<span>
-						{castLog.actionId!.name} from {castLog.timestamp.toFixed(2)}s to{' '}
-						{(castLog.castCancelledLog?.timestamp || castLog.timestamp + castLog.castTime).toFixed(2)}s{timingStr}
-						{travelTimeStr.length > 0 && travelTimeStr}
-					</span>
-					{totalDamage > 0 && (
-						<span>
-							Total: {totalDamage.toFixed(2)} ({(totalDamage / (castLog.effectiveTime || 1)).toFixed(2)} DPET)
-						</span>
-					)}
-					{castLog.damageDealtLogs.length > 0 && (
-						<ul className="rotation-timeline-cast-damage-list">
-							{castLog.damageDealtLogs.map(ddl => (
-								<li>
-									<span>
-										{ddl.timestamp.toFixed(2)}s - {ddl.result()}
-									</span>
-									{ddl.source?.isTarget && (
-										<span className="threat-metrics">
-											{' '}
-											({ddl.threat.toFixed(1)} {i18n.t('results_tab.details.timeline.tooltips.threat')})
-										</span>
-									)}
-								</li>
-							))}
-						</ul>
-					)}
-				</div>
-			);
-
-			const tooltip = tippy(castElem, {
-				placement: 'bottom',
-				content: tt,
-			});
-			this.addOnResetCallback(() => tooltip.destroy());
-
-			castLog.damageDealtLogs
-				.filter(ddl => ddl.tick)
-				.forEach(ddl => {
-					const tickElem = (
-						<div
-							className="rotation-timeline-tick"
-							style={{
-								left: this.timeToPx(ddl.timestamp),
-							}}
-						/>
-					);
-					rowElem.appendChild(tickElem);
-
-					const tt = (
-						<div className="timeline-tooltip">
-							<span>
-								{ddl.timestamp.toFixed(2)}s - {ddl.actionId!.name} {ddl.result()}
-							</span>
-							{ddl.source?.isTarget && (
-								<span className="threat-metrics">
-									{' '}
-									({ddl.threat.toFixed(1)} {i18n.t('results_tab.details.timeline.tooltips.threat')})
-								</span>
-							)}
-						</div>
-					);
-
-					const tooltip = tippy(tickElem, {
-						placement: 'bottom',
-						content: tt,
-					});
-					this.addOnResetCallback(() => tooltip.destroy());
-				});
-		});
-
-		// If there are any auras that correspond to this cast, visualize them in the same row.
-		aurasById
-			.filter(auraUptimeLogs => {
-				return idsToGroupForRotation.includes(actionId.spellId)
-					? actionId.equalsIgnoringTag(buffAuraToSpellIdMap[auraUptimeLogs[0].actionId!.spellId] ?? auraUptimeLogs[0].actionId!)
-					: actionId.equals(buffAuraToSpellIdMap[auraUptimeLogs[0].actionId!.spellId] ?? auraUptimeLogs[0].actionId!);
-			})
-			.forEach(auraUptimeLogs => this.applyAuraUptimeLogsToRow(auraUptimeLogs, rowElem, true));
-
-		this.rotationTimeline.appendChild(rowElem);
-	}
-
-	private addAuraRow(auraUptimeLogs: Array<AuraUptimeLog>, duration: number, scope = '') {
-		const actionId = auraUptimeLogs[0].actionId!;
-
-		const rowElem = this.makeRowElem(actionId, duration, scope);
-		this.rotationLabels.appendChild(this.makeLabelElem(actionId, false, true, scope));
-		this.rotationHiddenIdsContainer.appendChild(this.makeLabelElem(actionId, true, true, scope));
-		this.rotationTimeline.appendChild(rowElem);
-
-		this.applyAuraUptimeLogsToRow(auraUptimeLogs, rowElem, false);
-	}
-
-	private applyAuraUptimeLogsToRow(auraUptimeLogs: Array<AuraUptimeLog>, rowElem: JSX.Element, hasCast: boolean) {
-		auraUptimeLogs.forEach(aul => {
-			const auraElem = (
-				<div
-					className="rotation-timeline-aura"
-					style={{
-						left: this.timeToPx(aul.gainedAt),
-						minWidth: this.timeToPx(aul.fadedAt === aul.gainedAt ? 0.001 : aul.fadedAt - aul.gainedAt),
-					}}
-				/>
-			);
-			rowElem.appendChild(auraElem);
-
-			const tt = (
-				<div className="timeline-tooltip">
-					<span>
-						{aul.actionId!.name}: {aul.gainedAt.toFixed(2)}s - {aul.fadedAt.toFixed(2)}s
-					</span>
-				</div>
-			);
-
-			const tooltip = tippy(auraElem, {
-				placement: 'bottom',
-				content: tt,
-			});
-			this.addOnResetCallback(() => tooltip.destroy());
-
-			aul.stacksChange.forEach((scl, i) => {
-				if (scl.timestamp == aul.fadedAt) {
-					return;
-				}
-
-				const stacksChangeElem = (
-					<div
-						className="rotation-timeline-stacks-change"
-						style={{
-							left: this.timeToPx(scl.timestamp - aul.timestamp),
-							width: this.timeToPx(aul.stacksChange[i + 1] ? aul.stacksChange[i + 1].timestamp - scl.timestamp : aul.fadedAt - scl.timestamp),
-							textIndent: hasCast ? '30px' : undefined,
-						}}>
-						{String(scl.newStacks)}
-					</div>
-				);
-				auraElem.appendChild(stacksChangeElem);
-			});
-		});
-	}
-
-	private timeToPxValue(time: number): number {
-		return time * 100;
-	}
-	private timeToPx(time: number): string {
-		return this.timeToPxValue(time) + 'px';
-	}
-
-	private drawRotationTimeRuler(canvas: HTMLCanvasElement, duration: number) {
-		const height = 30;
-		canvas.width = this.timeToPxValue(duration);
-		canvas.height = height;
-
-		const ctx = canvas.getContext('2d')!;
-		ctx.strokeStyle = 'white';
-
-		ctx.font = 'bold 14px SimDefaultFont';
-		ctx.fillStyle = 'white';
-		ctx.lineWidth = 2;
-		ctx.beginPath();
-
-		// Bottom border line
-		ctx.moveTo(0, height);
-		ctx.lineTo(canvas.width, height);
-
-		// Tick lines
-		const numTicks = 1 + Math.floor(duration * 10);
-		for (let i = 0; i <= numTicks; i++) {
-			const time = i * 0.1;
-			let x = this.timeToPxValue(time);
-			if (i == 0) {
-				ctx.textAlign = 'left';
-				x++;
-			} else if (i % 10 == 0 && time + 1 > duration) {
-				ctx.textAlign = 'right';
-				x--;
-			} else {
-				ctx.textAlign = 'center';
-			}
-
-			let lineHeight = 0;
-			if (i % 10 == 0) {
-				lineHeight = height * 0.5;
-				ctx.fillText(time + 's', x, height - height * 0.6);
-			} else if (i % 5 == 0) {
-				lineHeight = height * 0.25;
-			} else {
-				lineHeight = height * 0.125;
-			}
-			ctx.moveTo(x, height);
-			ctx.lineTo(x, height - lineHeight);
-		}
-		ctx.stroke();
-	}
-
-	private dpsTooltip(log: DpsLog, _includeAuras: boolean, player: UnitMetrics, colorOverride: string) {
-		const showPlayerLabel = colorOverride != '';
-		return (
-			<div className="timeline-tooltip dps">
-				<div className="timeline-tooltip-header">
-					{showPlayerLabel ? (
-						<>
-							<img className="timeline-tooltip-icon" src={player.iconUrl} />
-							<span className="" style={{ color: colorOverride }}>
-								{player.label}
-							</span>
-							<span> - </span>
-						</>
-					) : null}
-					<span className="bold">{log.timestamp.toFixed(2)}s</span>
-				</div>
-				<div className="timeline-tooltip-body">
-					<ul className="timeline-dps-events">{log.damageLogs.map(damageLog => this.tooltipLogItem(damageLog, damageLog.result()))}</ul>
-					<div className="timeline-tooltip-body-row">
-						<span className="series-color">
-							{i18n.t('results_tab.details.timeline.tooltips.dps')}: {log.dps.toFixed(2)}
-						</span>
-					</div>
-				</div>
-				{this.tooltipAurasSection(log)}
-			</div>
-		);
-	}
-
-	private threatTooltip(log: ThreatLogGroup, includeAuras: boolean, player: UnitMetrics, colorOverride: string) {
-		const showPlayerLabel = colorOverride != '';
-		return (
-			<div className="timeline-tooltip threat">
-				<div className="timeline-tooltip-header">
-					{showPlayerLabel ? (
-						<>
-							<img className="timeline-tooltip-icon" src={player.iconUrl} />
-							<span className="" style={{ color: colorOverride }}>
-								{player.label}
-							</span>
-							<span> - </span>
-						</>
-					) : null}
-					<span className="bold">{log.timestamp.toFixed(2)}s</span>
-				</div>
-				<div className="timeline-tooltip-body">
-					<div className="timeline-tooltip-body-row">
-						<span className="series-color">
-							{i18n.t('results_tab.details.timeline.tooltips.before')}: {log.threatBefore.toFixed(1)}
-						</span>
-					</div>
-					<ul className="timeline-threat-events">
-						{log.logs.map(log =>
-							this.tooltipLogItem(
-								log,
-								<>
-									{log.threat.toFixed(1)} {i18n.t('results_tab.details.timeline.tooltips.threat')}
-								</>,
-							),
-						)}
-					</ul>
-					<div className="timeline-tooltip-body-row">
-						<span className="series-color">
-							{i18n.t('results_tab.details.timeline.tooltips.after')}: {log.threatAfter.toFixed(1)}
-						</span>
-					</div>
-				</div>
-				{includeAuras ? this.tooltipAurasSection(log) : null}
-			</div>
-		);
-	}
-
-	private resourceTooltipElem(log: ResourceChangedLogGroup, maxValue: number, includeAuras: boolean) {
-		const valToDisplayString = percentageResources.includes(log.resourceType)
-			? (val: number) => `${val.toFixed(1)} (${((val / maxValue) * 100).toFixed(0)}%)`
-			: (val: number) => `${val.toFixed(1)}`;
-
-		return (
-			<div className={`timeline-tooltip ${resourceNames.get(log.resourceType)!.toLowerCase().replaceAll(' ', '-')}`}>
-				<div className="timeline-tooltip-header">
-					<span className="bold">{log.timestamp.toFixed(2)}s</span>
-				</div>
-				<div className="timeline-tooltip-body">
-					<div className="timeline-tooltip-body-row">
-						<span className="series-color">
-							{i18n.t('results_tab.details.timeline.tooltips.before')}: {valToDisplayString(log.valueBefore)}
-						</span>
-					</div>
-					<ul className="timeline-mana-events">
-						{log.logs.map(manaChangedLog => this.tooltipLogItemElem(manaChangedLog, <>{manaChangedLog.resultString()}</>))}
-					</ul>
-					<div className="timeline-tooltip-body-row">
-						<span className="series-color">
-							{i18n.t('results_tab.details.timeline.tooltips.after')}: {valToDisplayString(log.valueAfter)}
-						</span>
-					</div>
-				</div>
-				{includeAuras && this.tooltipAurasSectionElem(log)}
-			</div>
-		);
-	}
-
-	private resourceTooltip(log: ResourceChangedLogGroup, maxValue: number, includeAuras: boolean) {
-		return this.resourceTooltipElem(log, maxValue, includeAuras);
-	}
-
-	private tooltipLogItem(log: SimLog, value: Element) {
-		return this.tooltipLogItemElem(log, value);
-	}
-
-	private tooltipLogItemElem(log: SimLog, value: Element): JSX.Element {
-		return (
-			<li>
-				{log.actionId && log.actionId.iconUrl && <img className="timeline-tooltip-icon" src={log.actionId.iconUrl}></img>}
-				{log.actionId && <span>{log.actionId.name}</span>}
-				<span className="series-color">{value}</span>
-			</li>
-		);
-	}
-
-	private tooltipAurasSection(log: SimLog) {
-		if (log.activeAuras.length == 0) {
-			return '';
-		}
-		return this.tooltipAurasSectionElem(log);
-	}
-
-	private tooltipAurasSectionElem(log: SimLog): JSX.Element {
-		if (log.activeAuras.length == 0) {
-			return <></>;
-		}
-
-		return (
-			<div className="timeline-tooltip-auras">
-				<div className="timeline-tooltip-body-row">
-					<span className="bold">{i18n.t('results_tab.details.timeline.tooltips.active_auras')}</span>
-				</div>
-				<ul className="timeline-active-auras">
-					{log.activeAuras.map(auraLog => (
-						<li>
-							{auraLog.actionId!.iconUrl && <img className="timeline-tooltip-icon" src={auraLog.actionId!.iconUrl}></img>}
-							<span>{auraLog.actionId!.name}</span>
-						</li>
-					))}
-				</ul>
-			</div>
-		);
-	}
-
-	update() {
-		if (!this.rendered) this.dpsResourcesPlot.render();
-		this.updatePlot();
-		this.rendered = true;
-	}
-
-	// Per-render resources (tooltips, listeners) belong to the slot being
-	// rendered so a parked slot can be destroyed independently. Rows are only
-	// ever built under a live slot (see updateRotationChart).
-	addOnResetCallback(callback: () => void) {
-		this.liveSlot!.resetCallbacks.push(callback);
-	}
-
-	// A single player's slot holds the rotation subtree plus every series, so the
-	// chart choice is irrelevant to it; multi-player options depend on the chart.
-	private resultKey(includeChart = false): string {
-		const rd = this.resultData!;
-		return [rd.result.request.requestId, JSON.stringify(rd.filter), includeChart ? this.chartPicker.value : ''].join('|');
-	}
-
-	// Single-player results offer the rotation chart; multi-player ones the threat chart.
-	private setRotationOptionVisible(visible: boolean) {
-		this.rootElem.querySelector('.rotation-option')!.classList.toggle('hide', !visible);
-		this.rootElem.querySelector('.threat-option')!.classList.toggle('hide', visible);
-	}
-
-	private static newSlot(key: string): RotationSlot {
-		return { key, labels: [], timeline: [], hiddenIdsNodes: [], emitter: new TypedEvent<void>(), resetCallbacks: [], plotOptions: null };
-	}
-
-	private takeParkedSlot(key: string): RotationSlot | null {
-		if (this.parkedSlot?.key !== key) return null;
-		const slot = this.parkedSlot;
-		this.parkedSlot = null;
-		return slot;
-	}
-
-	// Parks the on-screen subtree (nodes moved, tooltips kept alive), destroying
-	// the previously parked slot.
-	private stashLiveSlot() {
-		const slot = this.liveSlot;
-		if (!slot) return;
-		slot.labels = Array.from(this.rotationLabels.childNodes);
-		slot.timeline = Array.from(this.rotationTimeline.childNodes);
-		slot.hiddenIdsNodes = Array.from(this.rotationHiddenIdsContainer.childNodes);
-		this.rotationLabels.replaceChildren();
-		this.rotationTimeline.replaceChildren();
-		this.rotationHiddenIdsContainer.replaceChildren();
-		this.liveSlot = null;
-		if (this.parkedSlot) this.destroySlot(this.parkedSlot);
-		this.parkedSlot = slot;
-	}
-
-	private attachSlot(slot: RotationSlot) {
-		this.rotationLabels.replaceChildren(...slot.labels);
-		this.rotationTimeline.replaceChildren(...slot.timeline);
-		this.rotationHiddenIdsContainer.replaceChildren(...slot.hiddenIdsNodes);
-		this.liveSlot = slot;
-		// hiddenIds is global across results: re-apply it to the restored rows.
-		slot.emitter.emit(TypedEvent.nextEventID());
-	}
-
-	private destroySlot(slot: RotationSlot) {
-		slot.resetCallbacks.forEach(callback => callback());
-	}
-
-	render() {
-		if (this.rendered) return;
-		this.update();
-	}
-
-	reset() {
-		if (this.liveSlot) this.destroySlot(this.liveSlot);
-		if (this.parkedSlot) this.destroySlot(this.parkedSlot);
-		this.liveSlot = null;
-		this.parkedSlot = null;
-		super.reset();
-	}
-}
-
-const MELEE_ACTION_CATEGORY = 1;
-const SPELL_ACTION_CATEGORY = 2;
-const DEFAULT_ACTION_CATEGORY = 3;
-
-const auraAsResource: ActionId[] = [
-	ActionId.fromOtherId(OtherAction.OtherActionItemSwap, APLActionItemSwap_SwapSet.Main),
-	ActionId.fromOtherId(OtherAction.OtherActionItemSwap, APLActionItemSwap_SwapSet.Swap1),
-];
-
-// Hard-coded spell categories for controlling rotation ordering.
-const idToCategoryMap: Record<number, number> = {
-	[OtherAction.OtherActionMove]: 0,
-	[OtherAction.OtherActionItemSwap]: 0.015,
-	[OtherAction.OtherActionAttack]: 0.01,
-	[OtherAction.OtherActionShoot]: 0.5,
-
-	// Druid
-	[48480]: 0.1, // Maul
-	[48564]: MELEE_ACTION_CATEGORY + 0.1, // Mangle (Bear)
-	[48568]: MELEE_ACTION_CATEGORY + 0.2, // Lacerate
-	[48562]: MELEE_ACTION_CATEGORY + 0.3, // Swipe (Bear)
-
-	[48566]: MELEE_ACTION_CATEGORY + 0.1, // Mangle (Cat)
-	[48572]: MELEE_ACTION_CATEGORY + 0.2, // Shred
-	[49800]: MELEE_ACTION_CATEGORY + 0.51, // Rip
-	[52610]: MELEE_ACTION_CATEGORY + 0.52, // Savage Roar
-	[48577]: MELEE_ACTION_CATEGORY + 0.53, // Ferocious Bite
-
-	[48465]: SPELL_ACTION_CATEGORY + 0.1, // Starfire
-	[48461]: SPELL_ACTION_CATEGORY + 0.2, // Wrath
-	[53201]: SPELL_ACTION_CATEGORY + 0.3, // Starfall
-	[48463]: SPELL_ACTION_CATEGORY + 0.4, // Moonfire
-
-	// Hunter
-	[27014]: 0.02, // Raptor Strike
-	[32120]: MELEE_ACTION_CATEGORY + 0.2, // Steady Shot
-	[27065]: MELEE_ACTION_CATEGORY + 0.21, // Aimed Shot
-	[27021]: MELEE_ACTION_CATEGORY + 0.22, // Multi Shot
-	[27019]: MELEE_ACTION_CATEGORY + 0.23, // Arcane Shot
-	[34026]: MELEE_ACTION_CATEGORY + 0.24, // Kill Command
-	[27016]: MELEE_ACTION_CATEGORY + 0.3, // Serpent Sting
-	[3043]: MELEE_ACTION_CATEGORY + 0.31, // Scorpid Sting
-
-	// Paladin
-	[31893]: MELEE_ACTION_CATEGORY + 0.001, // Seal of Blood proc
-	[20424]: MELEE_ACTION_CATEGORY + 0.002, // Seal of Command proc
-	[25742]: MELEE_ACTION_CATEGORY + 0.003, // Seal of Righteousness proc (Rank 1)
-	[25740]: MELEE_ACTION_CATEGORY + 0.0031, // Seal of Righteousness proc (Rank 2)
-	[25739]: MELEE_ACTION_CATEGORY + 0.0032, // Seal of Righteousness proc (Rank 3)
-	[25738]: MELEE_ACTION_CATEGORY + 0.0033, // Seal of Righteousness proc (Rank 4)
-	[25737]: MELEE_ACTION_CATEGORY + 0.0034, // Seal of Righteousness proc (Rank 5)
-	[25736]: MELEE_ACTION_CATEGORY + 0.0035, // Seal of Righteousness proc (Rank 6)
-	[25735]: MELEE_ACTION_CATEGORY + 0.0036, // Seal of Righteousness proc (Rank 7)
-	[25713]: MELEE_ACTION_CATEGORY + 0.0037, // Seal of Righteousness proc (Rank 8)
-	[27156]: MELEE_ACTION_CATEGORY + 0.0038, // Seal of Righteousness proc (Rank 9)
-	[42463]: MELEE_ACTION_CATEGORY + 0.004, // Seal of Vengeance proc
-	[31803]: MELEE_ACTION_CATEGORY + 0.0041, // Holy Vengeance
-	[20167]: MELEE_ACTION_CATEGORY + 0.005, // Seal of Light proc (Rank 1)
-	[20333]: MELEE_ACTION_CATEGORY + 0.0051, // Seal of Light proc (Rank 2)
-	[20334]: MELEE_ACTION_CATEGORY + 0.0052, // Seal of Light proc (Rank 3)
-	[20340]: MELEE_ACTION_CATEGORY + 0.0053, // Seal of Light proc (Rank 4)
-	[27161]: MELEE_ACTION_CATEGORY + 0.0054, // Seal of Light proc (Rank 5)
-	[20168]: MELEE_ACTION_CATEGORY + 0.006, // Seal of Wisdom proc (Rank 1)
-	[20350]: MELEE_ACTION_CATEGORY + 0.0061, // Seal of Wisdom proc (Rank 2)
-	[20351]: MELEE_ACTION_CATEGORY + 0.0062, // Seal of Wisdom proc (Rank 3)
-	[27167]: MELEE_ACTION_CATEGORY + 0.0063, // Seal of Wisdom proc (Rank 4)
-	[20170]: MELEE_ACTION_CATEGORY + 0.007, // Seal of Justice proc
-	[31892]: MELEE_ACTION_CATEGORY + 0.02, // Seal of Blood
-	[20375]: MELEE_ACTION_CATEGORY + 0.021, // Seal of Command (Rank 1)
-	[20915]: MELEE_ACTION_CATEGORY + 0.0211, // Seal of Command (Rank 2)
-	[20918]: MELEE_ACTION_CATEGORY + 0.0212, // Seal of Command (Rank 3)
-	[20919]: MELEE_ACTION_CATEGORY + 0.0213, // Seal of Command (Rank 4)
-	[20920]: MELEE_ACTION_CATEGORY + 0.0214, // Seal of Command (Rank 5)
-	[27170]: MELEE_ACTION_CATEGORY + 0.0215, // Seal of Command (Rank 6)
-	[21084]: MELEE_ACTION_CATEGORY + 0.022, // Seal of Righteousness (Rank 1)
-	[20287]: MELEE_ACTION_CATEGORY + 0.0221, // Seal of Righteousness (Rank 2)
-	[20288]: MELEE_ACTION_CATEGORY + 0.0222, // Seal of Righteousness (Rank 3)
-	[20289]: MELEE_ACTION_CATEGORY + 0.0223, // Seal of Righteousness (Rank 4)
-	[20290]: MELEE_ACTION_CATEGORY + 0.0224, // Seal of Righteousness (Rank 5)
-	[20291]: MELEE_ACTION_CATEGORY + 0.0225, // Seal of Righteousness (Rank 6)
-	[20292]: MELEE_ACTION_CATEGORY + 0.0226, // Seal of Righteousness (Rank 7)
-	[20293]: MELEE_ACTION_CATEGORY + 0.0227, // Seal of Righteousness (Rank 8)
-	[27155]: MELEE_ACTION_CATEGORY + 0.0228, // Seal of Righteousness (Rank 9)
-	[31801]: MELEE_ACTION_CATEGORY + 0.0229, // Seal of Vengeance
-	[20165]: MELEE_ACTION_CATEGORY + 0.02291, // Seal of Light (Rank 1)
-	[20347]: MELEE_ACTION_CATEGORY + 0.022911, // Seal of Light (Rank 2)
-	[20348]: MELEE_ACTION_CATEGORY + 0.022912, // Seal of Light (Rank 3)
-	[20349]: MELEE_ACTION_CATEGORY + 0.022913, // Seal of Light (Rank 4)
-	[27160]: MELEE_ACTION_CATEGORY + 0.022914, // Seal of Light (Rank 5)
-	[20166]: MELEE_ACTION_CATEGORY + 0.02292, // Seal of Wisdom (Rank 1)
-	[20356]: MELEE_ACTION_CATEGORY + 0.022921, // Seal of Wisdom (Rank 2)
-	[20357]: MELEE_ACTION_CATEGORY + 0.022922, // Seal of Wisdom (Rank 3)
-	[27166]: MELEE_ACTION_CATEGORY + 0.022923, // Seal of Wisdom (Rank 4)
-	[20164]: MELEE_ACTION_CATEGORY + 0.02293, // Seal of Justice (Rank 1)
-	[31895]: MELEE_ACTION_CATEGORY + 0.022931, // Seal of Justice (Rank 2)
-	[20271]: MELEE_ACTION_CATEGORY + 0.03, // Judgement
-	[31898]: MELEE_ACTION_CATEGORY + 0.04, // Judgement of Blood
-	[20425]: MELEE_ACTION_CATEGORY + 0.041, // Judgement of Command (Rank 1)
-	[20962]: MELEE_ACTION_CATEGORY + 0.0411, // Judgement of Command (Rank 2)
-	[20961]: MELEE_ACTION_CATEGORY + 0.0412, // Judgement of Command (Rank 3)
-	[20967]: MELEE_ACTION_CATEGORY + 0.0413, // Judgement of Command (Rank 4)
-	[20968]: MELEE_ACTION_CATEGORY + 0.0414, // Judgement of Command (Rank 5)
-	[27172]: MELEE_ACTION_CATEGORY + 0.0415, // Judgement of Command (Rank 6)
-	[20187]: MELEE_ACTION_CATEGORY + 0.042, // Judgement of Righteousness (Rank 1)
-	[20280]: MELEE_ACTION_CATEGORY + 0.0421, // Judgement of Righteousness (Rank 2)
-	[20281]: MELEE_ACTION_CATEGORY + 0.0422, // Judgement of Righteousness (Rank 3)
-	[20282]: MELEE_ACTION_CATEGORY + 0.0423, // Judgement of Righteousness (Rank 4)
-	[20283]: MELEE_ACTION_CATEGORY + 0.0424, // Judgement of Righteousness (Rank 5)
-	[20284]: MELEE_ACTION_CATEGORY + 0.0425, // Judgement of Righteousness (Rank 6)
-	[20285]: MELEE_ACTION_CATEGORY + 0.0426, // Judgement of Righteousness (Rank 7)
-	[20286]: MELEE_ACTION_CATEGORY + 0.0427, // Judgement of Righteousness (Rank 8)
-	[27157]: MELEE_ACTION_CATEGORY + 0.0428, // Judgement of Righteousness (Rank 9)
-	[31804]: MELEE_ACTION_CATEGORY + 0.043, // Judgement of Vengeance
-	[21183]: MELEE_ACTION_CATEGORY + 0.044, // Judgement of the Crusader (Rank 1)
-	[20188]: MELEE_ACTION_CATEGORY + 0.0441, // Judgement of the Crusader (Rank 2)
-	[20300]: MELEE_ACTION_CATEGORY + 0.0442, // Judgement of the Crusader (Rank 3)
-	[20301]: MELEE_ACTION_CATEGORY + 0.0443, // Judgement of the Crusader (Rank 4)
-	[20302]: MELEE_ACTION_CATEGORY + 0.0444, // Judgement of the Crusader (Rank 5)
-	[20303]: MELEE_ACTION_CATEGORY + 0.0445, // Judgement of the Crusader (Rank 6)
-	[27159]: MELEE_ACTION_CATEGORY + 0.0446, // Judgement of the Crusader (Rank 7)
-	[20185]: MELEE_ACTION_CATEGORY + 0.045, // Judgement of Light (Rank 1)
-	[20344]: MELEE_ACTION_CATEGORY + 0.0451, // Judgement of Light (Rank 2)
-	[20345]: MELEE_ACTION_CATEGORY + 0.0452, // Judgement of Light (Rank 3)
-	[20346]: MELEE_ACTION_CATEGORY + 0.0453, // Judgement of Light (Rank 4)
-	[27162]: MELEE_ACTION_CATEGORY + 0.0454, // Judgement of Light (Rank 5)
-	[20186]: MELEE_ACTION_CATEGORY + 0.046, // Judgement of Wisdom (Rank 1)
-	[20354]: MELEE_ACTION_CATEGORY + 0.0461, // Judgement of Wisdom (Rank 2)
-	[20355]: MELEE_ACTION_CATEGORY + 0.0462, // Judgement of Wisdom (Rank 3)
-	[27164]: MELEE_ACTION_CATEGORY + 0.0463, // Judgement of Wisdom (Rank 4)
-	[20184]: MELEE_ACTION_CATEGORY + 0.047, // Judgement of Justice (Rank 1)
-	[31896]: MELEE_ACTION_CATEGORY + 0.0471, // Judgement of Justice (Rank 2)
-	[35395]: MELEE_ACTION_CATEGORY + 0.05, // Crusader Strike
-	[31935]: MELEE_ACTION_CATEGORY + 0.05, // Avenger's Shield (Rank 1)
-	[32699]: MELEE_ACTION_CATEGORY + 0.051, // Avenger's Shield (Rank 2)
-	[32700]: MELEE_ACTION_CATEGORY + 0.052, // Avenger's Shield (Rank 3)
-	[879]: MELEE_ACTION_CATEGORY + 0.06, // Exorcism (Rank 1)
-	[5614]: MELEE_ACTION_CATEGORY + 0.061, // Exorcism (Rank 2)
-	[5615]: MELEE_ACTION_CATEGORY + 0.062, // Exorcism (Rank 3)
-	[10312]: MELEE_ACTION_CATEGORY + 0.063, // Exorcism (Rank 4)
-	[10313]: MELEE_ACTION_CATEGORY + 0.064, // Exorcism (Rank 5)
-	[10314]: MELEE_ACTION_CATEGORY + 0.065, // Exorcism (Rank 6)
-	[27138]: MELEE_ACTION_CATEGORY + 0.066, // Exorcism (Rank 7)
-	[26573]: MELEE_ACTION_CATEGORY + 0.07, // Consecration (Rank 1)
-	[20116]: MELEE_ACTION_CATEGORY + 0.071, // Consecration (Rank 2)
-	[20922]: MELEE_ACTION_CATEGORY + 0.072, // Consecration (Rank 3)
-	[20923]: MELEE_ACTION_CATEGORY + 0.073, // Consecration (Rank 4)
-	[20924]: MELEE_ACTION_CATEGORY + 0.074, // Consecration (Rank 5)
-	[27173]: MELEE_ACTION_CATEGORY + 0.075, // Consecration (Rank 6)
-	[2812]: MELEE_ACTION_CATEGORY + 0.08, // Holy Wrath (Rank 1)
-	[10318]: MELEE_ACTION_CATEGORY + 0.081, // Holy Wrath (Rank 2)
-	[27139]: MELEE_ACTION_CATEGORY + 0.082, // Holy Wrath (Rank 3)
-	[24275]: MELEE_ACTION_CATEGORY + 0.09, // Hammer of Wrath (Rank 1)
-	[24274]: MELEE_ACTION_CATEGORY + 0.091, // Hammer of Wrath (Rank 2)
-	[24239]: MELEE_ACTION_CATEGORY + 0.092, // Hammer of Wrath (Rank 3)
-	[27180]: MELEE_ACTION_CATEGORY + 0.093, // Hammer of Wrath (Rank 4)
-	[31884]: SPELL_ACTION_CATEGORY + 0.01, // Avenging Wrath
-	[25780]: SPELL_ACTION_CATEGORY + 0.94, // Righteous Fury
-	[27149]: SPELL_ACTION_CATEGORY + 0.941, // Devotion Aura
-	[19746]: SPELL_ACTION_CATEGORY + 0.942, // Concentration Aura
-	[27153]: SPELL_ACTION_CATEGORY + 0.943, // Fire Resistance Aura
-	[27152]: SPELL_ACTION_CATEGORY + 0.944, // Frost Resistance Aura
-	[27151]: SPELL_ACTION_CATEGORY + 0.945, // Shadow Resistance Aura
-	[20218]: SPELL_ACTION_CATEGORY + 0.946, // Sanctity Aura
-
-	// Priest
-	[48300]: SPELL_ACTION_CATEGORY + 0.11, // Devouring Plague
-	[48125]: SPELL_ACTION_CATEGORY + 0.12, // Shadow Word: Pain
-	[48160]: SPELL_ACTION_CATEGORY + 0.13, // Vampiric Touch
-	[48135]: SPELL_ACTION_CATEGORY + 0.14, // Holy Fire
-	[48123]: SPELL_ACTION_CATEGORY + 0.19, // Smite
-	[48127]: SPELL_ACTION_CATEGORY + 0.2, // Mind Blast
-	[48158]: SPELL_ACTION_CATEGORY + 0.3, // Shadow Word: Death
-	[48156]: SPELL_ACTION_CATEGORY + 0.4, // Mind Flay
-
-	// Rogue
-	[6774]: MELEE_ACTION_CATEGORY + 0.1, // Slice and Dice
-	[8647]: MELEE_ACTION_CATEGORY + 0.2, // Expose Armor
-	[48672]: MELEE_ACTION_CATEGORY + 0.3, // Rupture
-	[57993]: MELEE_ACTION_CATEGORY + 0.3, // Envenom
-	[48668]: MELEE_ACTION_CATEGORY + 0.4, // Eviscerate
-	[48666]: MELEE_ACTION_CATEGORY + 0.5, // Mutilate
-	[48665]: MELEE_ACTION_CATEGORY + 0.6, // Mutilate (MH)
-	[48664]: MELEE_ACTION_CATEGORY + 0.7, // Mutilate (OH)
-	[48638]: MELEE_ACTION_CATEGORY + 0.5, // Sinister Strike
-	[51723]: MELEE_ACTION_CATEGORY + 0.8, // Fan of Knives
-	[57973]: SPELL_ACTION_CATEGORY + 0.1, // Deadly Poison
-	[57968]: SPELL_ACTION_CATEGORY + 0.2, // Instant Poison
-
-	// Shaman
-	[8232]: 0.11, // Windfury Weapon
-	[8024]: 0.12, // Flametongue Weapon
-	[8033]: 0.12, // Frostbrand Weapon
-	[17364]: MELEE_ACTION_CATEGORY + 0.1, // Stormstrike
-	[60103]: MELEE_ACTION_CATEGORY + 0.2, // Lava Lash
-	[49233]: SPELL_ACTION_CATEGORY + 0.21, // Flame Shock
-	[49231]: SPELL_ACTION_CATEGORY + 0.22, // Earth Shock
-	[49236]: SPELL_ACTION_CATEGORY + 0.23, // Frost Shock
-	[60043]: SPELL_ACTION_CATEGORY + 0.31, // Lava Burst
-	[49238]: SPELL_ACTION_CATEGORY + 0.32, // Lightning Bolt
-	[49271]: SPELL_ACTION_CATEGORY + 0.33, // Chain Lightning
-	[61657]: SPELL_ACTION_CATEGORY + 0.41, // Fire Nova
-	[58734]: SPELL_ACTION_CATEGORY + 0.42, // Magma Totem
-	[58704]: SPELL_ACTION_CATEGORY + 0.43, // Searing Totem
-	[49281]: SPELL_ACTION_CATEGORY + 0.51, // Lightning Shield
-	[49279]: SPELL_ACTION_CATEGORY + 0.52, // Lightning Shield (Proc)
-	[2825]: DEFAULT_ACTION_CATEGORY + 0.1, // Bloodlust
-
-	// Warlock
-	[30910]: SPELL_ACTION_CATEGORY + 0.02, // Curse of Doom
-	[27218]: SPELL_ACTION_CATEGORY + 0.02, // Curse of Agony
-	[27226]: SPELL_ACTION_CATEGORY + 0.02, // Curse of Recklessness
-	[27228]: SPELL_ACTION_CATEGORY + 0.02, // Curse of Elements
-	[172]: SPELL_ACTION_CATEGORY + 0.1, // Corruption
-	[48181]: SPELL_ACTION_CATEGORY + 0.2, // Haunt
-	[30108]: SPELL_ACTION_CATEGORY + 0.3, // Unstable Affliction
-	[348]: SPELL_ACTION_CATEGORY + 0.31, // Immolate
-	[30912]: SPELL_ACTION_CATEGORY + 0.32, // Conflagrate
-	[686]: SPELL_ACTION_CATEGORY + 0.5, // Shadow Bolt
-	[29722]: SPELL_ACTION_CATEGORY + 0.51, // Incinerate
-	[1120]: SPELL_ACTION_CATEGORY + 0.6, // Drain Soul
-	[1454]: SPELL_ACTION_CATEGORY + 0.7, // Life Tap
-
-	// Mage
-	[42842]: SPELL_ACTION_CATEGORY + 0.01, // Frostbolt
-	[42897]: SPELL_ACTION_CATEGORY + 0.02, // Arcane Blast
-	[42833]: SPELL_ACTION_CATEGORY + 0.02, // Fireball
-	[10]: SPELL_ACTION_CATEGORY + 0.021, // Blizzard - Cast
-	[42208]: SPELL_ACTION_CATEGORY + 0.022, // Blizzard - Tick
-	[42859]: SPELL_ACTION_CATEGORY + 0.03, // Scorch
-	[42891]: SPELL_ACTION_CATEGORY + 0.1, // Pyroblast
-	[42846]: SPELL_ACTION_CATEGORY + 0.1, // Arcane Missiles
-	[44572]: SPELL_ACTION_CATEGORY + 0.1, // Deep Freeze
-	[44781]: SPELL_ACTION_CATEGORY + 0.2, // Arcane Barrage
-	[42914]: SPELL_ACTION_CATEGORY + 0.2, // Ice Lance
-	[12654]: SPELL_ACTION_CATEGORY + 0.3, // Ignite
-	[12472]: SPELL_ACTION_CATEGORY + 0.4, // Icy Veins
-	[11129]: SPELL_ACTION_CATEGORY + 0.4, // Combustion
-	[12042]: SPELL_ACTION_CATEGORY + 0.4, // Arcane Power
-	[11958]: SPELL_ACTION_CATEGORY + 0.41, // Cold Snap
-	[12043]: SPELL_ACTION_CATEGORY + 0.41, // Presence of Mind
-	[31687]: SPELL_ACTION_CATEGORY + 0.41, // Water Elemental
-	[55342]: SPELL_ACTION_CATEGORY + 0.5, // Mirror Image
-	[33312]: SPELL_ACTION_CATEGORY + 0.51, // Mana Gems
-	[12051]: SPELL_ACTION_CATEGORY + 0.52, // Evocate
-	[44448]: SPELL_ACTION_CATEGORY + 0.6, // Hot Streak
-	[12536]: SPELL_ACTION_CATEGORY + 0.61, // Clearcasting
-
-	// Warrior
-	[25231]: 0.01 - 0.001, // Cleave
-	[30324]: 0.01 - 0.001, // Heroic Strike
-	[12723]: SPELL_ACTION_CATEGORY + 0.04, // Sweeping Strikes
-	[25242]: MELEE_ACTION_CATEGORY + 0.05, // Slam
-	[30335]: MELEE_ACTION_CATEGORY + 0.1, // Bloodthirst
-	[30330]: MELEE_ACTION_CATEGORY + 0.1, // Mortal Strike
-	[30356]: MELEE_ACTION_CATEGORY + 0.1, // Shield Slam
-	[30022]: MELEE_ACTION_CATEGORY + 0.21, // Devastate
-	[25225]: MELEE_ACTION_CATEGORY + 0.211, // Sunder Armor
-	[30357]: MELEE_ACTION_CATEGORY + 0.23, // Revenge
-	[1680]: MELEE_ACTION_CATEGORY + 0.24, // Whirlwind
-	[11585]: MELEE_ACTION_CATEGORY + 0.25, // Overpower
-	[25236]: MELEE_ACTION_CATEGORY + 0.42, // Execute
-	[12867]: SPELL_ACTION_CATEGORY + 0.51, // Deep Wounds
-	[2565]: SPELL_ACTION_CATEGORY + 0.62, // Shield Block
-	[12292]: SPELL_ACTION_CATEGORY + 0.63, // CD - Death Wish
-	[1719]: SPELL_ACTION_CATEGORY + 0.63, // CD - Recklnessness
-	[30033]: SPELL_ACTION_CATEGORY + 0.63, // CD - Rampage
-	[871]: SPELL_ACTION_CATEGORY + 0.63, // CD - Shield Wall
-	[20230]: SPELL_ACTION_CATEGORY + 0.63, // CD - Retaliation
-	[2687]: SPELL_ACTION_CATEGORY + 0.63, // CD - Bloodrage
-	[18499]: SPELL_ACTION_CATEGORY + 0.63, // CD - Berserker Rage
-	[12975]: SPELL_ACTION_CATEGORY + 0.63, // CD - Last Stand
-
-	[71]: DEFAULT_ACTION_CATEGORY + 0.1, // Defensive Stance
-	[2457]: DEFAULT_ACTION_CATEGORY + 0.1, // Battle Stance
-	[2458]: DEFAULT_ACTION_CATEGORY + 0.1, // Berserker Stance
-	[2048]: DEFAULT_ACTION_CATEGORY + 0.1, // Battle Shout
-	[469]: DEFAULT_ACTION_CATEGORY + 0.1, // Commanding Shout
-
-	// Generic
-	[20925]: SPELL_ACTION_CATEGORY + 0.931, // Holy Shield (Rank 1)
-	[20927]: SPELL_ACTION_CATEGORY + 0.931, // Holy Shield (Rank 2)
-	[20928]: SPELL_ACTION_CATEGORY + 0.931, // Holy Shield (Rank 3)
-	[27179]: SPELL_ACTION_CATEGORY + 0.931, // Holy Shield (Rank 4)
-	[27150]: SPELL_ACTION_CATEGORY + 0.932, // Retribution Aura
-	[27169]: SPELL_ACTION_CATEGORY + 0.933, // Blessing of Sanctuary
-	[26992]: SPELL_ACTION_CATEGORY + 0.934, // Thorns
-	[23827]: SPELL_ACTION_CATEGORY + 0.941, // Super Sapper
-	[10646]: SPELL_ACTION_CATEGORY + 0.941, // Goblin Sapper
-	[23736]: SPELL_ACTION_CATEGORY + 0.942, // Fel Iron Bomb
-	[23737]: SPELL_ACTION_CATEGORY + 0.943, // Adamantite Grenade
-	[23841]: SPELL_ACTION_CATEGORY + 0.944, // Gnomish Flame Turret
-};
-
-const idsToGroupForRotation: Array<number> = [
-	6774, // Rogue - Slice and Dice
-	26865, // Rogue - Eviscerate
-	26867, // Rogue - Rupture
-	32684, // Rogue - Envenom
-	26866, // Rogue - Expose Armor
-];
-
-const percentageResources: Array<ResourceType> = [ResourceType.ResourceTypeHealth, ResourceType.ResourceTypeMana];
