+++ ui/core/components/detailed_results/timeline/index.ts
@@ -1,0 +1,1 @@
+export { Timeline } from './timeline';
