+++ ui/core/components/detailed_results/timeline/types.ts
@@ -1,0 +1,19 @@
+import { TypedEvent } from '../../../typed_event';
+import { ResultComponentConfig } from '../result_component';
+import { WindowedRow } from './windowed_row';
+
+export type TooltipHandler = (dataPointIndex: number) => Element;
+
+export type TimelineConfig = ResultComponentConfig;
+
+export interface RotationSlot {
+	key: string;
+	labels: Array<Node>;
+	timeline: Array<Node>;
+	hiddenIdsNodes: Array<Node>;
+	emitter: TypedEvent<void>;
+	resetCallbacks: Array<() => void>;
+	plotOptions: any;
+	// Rows whose contents are populated from the horizontal scroll position.
+	windowedRows: Array<WindowedRow>;
+}
