+++ ui/scss/core/components/detailed_results/_timeline.scss
@@ -77,7 +77,25 @@
 	}
 
 	.timeline-chart-picker {
-		width: 8rem;
+		width: auto;
+		margin-left: auto;
+		flex-shrink: 0;
+		align-self: flex-start;
+
+		.btn-outline-primary {
+			--bs-btn-hover-bg: var(--bs-primary);
+			--bs-btn-hover-border-color: var(--bs-primary);
+			--bs-btn-active-bg: var(--bs-primary);
+			--bs-btn-active-border-color: var(--bs-primary);
+		}
+
+		.btn-check:focus + .btn {
+			box-shadow: none;
+		}
+
+		.btn-check:focus-visible + .btn {
+			box-shadow: 0 0 0 0.25rem rgba(var(--bs-primary-rgb), 0.5);
+		}
 	}
 
 	.timeline-show-gcd {
@@ -104,8 +122,6 @@
 	margin: auto;
 }
 
-// The timeline is generated by ApexCharts which sets height manually. In order to scale it properly, we need to apply several
-// height: 100% attributes on these containers
 .timeline-content {
 	height: 100%;
 
@@ -132,6 +148,9 @@
 		border: var(--bs-tooltip-border) 1px solid;
 		color: var(--bs-tooltip-body-color);
 		padding: var(--bs-tooltip-body-padding-y) var(--bs-tooltip-body-padding-x);
+		max-width: min(94vw, 68rem);
+		max-height: 80vh;
+		overflow: hidden;
 	}
 }
 
@@ -139,6 +158,9 @@
 	display: flex;
 	flex-direction: column;
 	gap: var(--spacer-3);
+	max-height: 75vh;
+	max-width: min(92vw, 64rem);
+	overflow: auto;
 
 	ul {
 		display: flex;
@@ -155,6 +177,23 @@
 		align-items: center;
 		gap: var(--spacer-2);
 	}
+}
+
+.timeline-tooltip .timeline-active-auras {
+	display: block;
+	max-height: 56vh;
+	column-count: 2;
+	column-gap: var(--spacer-3);
+
+	li {
+		break-inside: avoid;
+		margin-bottom: var(--spacer-1);
+	}
+}
+
+.timeline-tooltip .timeline-dps-events {
+	max-height: 40vh;
+	overflow-y: auto;
 }
 
 .timeline-tooltip-header {
